% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdisagg-package.R
\docType{data}
\name{gdp.q}
\alias{gdp.q}
\title{Gross Domestic Product}
\format{
A \code{data.frame}.
}
\source{
State Secretariat for Economic Affairs (SECO).
}
\description{
Qarterly real GDP, not seasonally adjusted, in millions of Swiss Francs
(2010 prices).
}
\examples{
\dontrun{
# recreate the series with latest data
library(tsbox)
library(dplyr)
library(dataseries)
library(imputeTS)
dataseries::ds("ch_seco_gdp.nsa.real.gdp") \%>\%
  ts_default() \%>\%
  ts_span(start = 2005)
}
}
\keyword{datasets}
