% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdisagg-package.R
\docType{data}
\name{spi.d}
\alias{spi.d}
\title{SPI Swiss Performance Index}
\format{
A \code{data.frame}.
}
\source{
Swiss National Bank (SNB)
}
\description{
Daily values of stock market index.
}
\examples{
\dontrun{
# recreate the series with latest data
library(tsbox)
library(dplyr)
library(dataseries)
library(imputeTS)
dataseries::ds("ch_snb_capchstocki.gdr") \%>\%
  ts_default() \%>\%
  ts_regular() \%>\%
  imputeTS::na_interpolation(option = "spline") \%>\%
  ts_span(start = 2005)
}
}
\keyword{datasets}
