\name{swisspharma}
\docType{data}
\alias{exports.m}
\alias{exports.q}
\alias{imports.q}
\alias{sales.a}
\alias{sales.q}
\title{Trade and Sales of Chemical and Pharmaceutical Industry}
\description{
This data set contains the monthly and quarterly imports and exports of the
chemical and pharmaceutical industry in Switzerland (in Mio. Swiss Francs) as
well as their quarterly and annual sales (Index).
}
\usage{exports.m
exports.q
imports.q
sales.a
sales.q
}
\format{Each time series is an object of class \code{"ts"}. The number of observations
depends on the frequency.}
\source{
Import and Export Data are from the Swiss Federal Customs Administration. Sales
Data are from the Swiss Federal Statistical Office.
}
\keyword{datasets}
