\name{tempdisagg-internal}
\alias{SubConvertEnd}
\alias{SubConvertStart}
\alias{SubAggregation}
\alias{CalcGLS}
\alias{CalcC}
\alias{CalcPowerMatrix}
\alias{CalcQ}
\alias{CalcQ_Lit}
\alias{CalcR}
\alias{CalcS}
\alias{SubDenton}
\alias{SubRegressionBased}


\title{Internal tempdisagg functions}
\description{
  Internal functions
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
