\docType{package}
\name{tempdisagg-package}
\alias{tempdisagg}
\alias{tempdisagg-package}
\title{Methods for Temporal Disaggregation and Interpolation of Time Series}
\description{
  Temporal disaggregation methods are used to disaggregate
  and interpolate a low frequency time series to a higher
  frequency series. This can be done without additional
  information or by exploiting the information contained in
  one or more indicators. All disaggregation methods ensure
  that either the sum, the average, the first or the last
  value of the resulting high frequency series is
  consistent with the low frequency series.

  The best way to start is to run the interactive demo:

  \code{demo(tempdisagg)}
}
\author{
  Christoph Sax \email{christoph.sax@gmail.com}, Peter
  Steiner
}
\seealso{
  \code{\link{td}} for more information on usage.
}
\keyword{package}

